requirements.txt
streamlit==1.32.0
boto3
python-dotenv
